refactor(hub): move products component
