+++ packages/manager/modules/hub/src/components/products/products.component.js
