+++ packages/manager/modules/hub/src/components/products/products.html
