+++ packages/manager/modules/hub/src/components/products/translations/Messages_fr_FR.json
