+++ packages/manager/modules/hub/src/index.js
@@ -5,10 +5,10 @@ import '@ovh-ux/ng-ovh-payment-method';
 import '@uirouter/angularjs';
 import 'angular-translate';
 
-import ovhManagerHubProducts from './products';
-import ovhManagerHubCarousel from './components/carousel';
-import ovhManagerHubTile from './components/tile';
 import billingSummary from './components/billing-summary';
+import ovhManagerHubCarousel from './components/carousel';
+import ovhManagerHubProducts from './components/products';
+import ovhManagerHubTile from './components/tile';
 import userPanel from './components/user-panel';
 
 const moduleName = 'ovhManagerHub';
